+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 0.2.0 - Wednesday, February 23, 2022
+- Updated to the latest React Spectrum release ([15 February 2022](https://react-spectrum.adobe.com/releases/2022-02-15.html))
+- Added support for the ContextualHelp and ActionMenu components
+- Added support for the 'useProvider' hook
+- Removed hacks from the ToggleButton and NumberField components that are no longer needed
+
 ### 0.1.1 - Tuesday, January 18, 2022
 - Added support for the the Provider.breakpoints prop